consolidate patch code
